layout changes, a couple of filter bugfixes to the cpp

+++ app/src/main/cpp/AudioProcessors/Filters/ButterworthLowPassFilter.cpp
@@ -13,16 +13,16 @@ namespace wavetablesynthesizer {
         const double pi = M_PI;
         const double qFactor = M_SQRT1_2;
 
-        auto normalisedCutoff = 2 * pi * (frequencyCutoff / _sampleRate);
+        auto normalisedCutoff = 2 * pi * (static_cast<double>(frequencyCutoff) / _sampleRate);
         auto cutoffTangent = tan(normalisedCutoff / 2);
         auto cutoffTangentSquared = pow(cutoffTangent, 2.0);
         auto alpha = 1 + (cutoffTangent / qFactor) + cutoffTangentSquared;
 
-        _a1 = 2 * (cutoffTangentSquared - 1) / alpha;
-        _a2 = (1 - (cutoffTangent / qFactor) + cutoffTangentSquared) / alpha;
+        _a1 = (float)(2 * (cutoffTangentSquared - 1) / alpha);
+        _a2 = (float)((1 - (cutoffTangent / qFactor) + cutoffTangentSquared) / alpha);
 
-        _b0 = cutoffTangentSquared / alpha;
-        _b1 = 2 * _b0;
+        _b0 = (float)(cutoffTangentSquared / alpha);
+        _b1 = (float)(2 * _b0);
         _b2 = _b0;
     }
 }
